Split view model
